+++ Source/Bugsnag.m
@@ -220,6 +220,17 @@ static NSMutableArray <id<BugsnagPlugin>> *registeredPlugins;
     }
 }
 
++ (void)leaveBreadcrumbWithMessage:(NSString *_Nonnull)message
+                       andMetadata:(NSDictionary *_Nullable)metadata
+{
+    if ([self bugsnagStarted]) {
+        [self leaveBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumbs) {
+            crumbs.message = message;
+            crumbs.metadata = metadata;
+        }];
+    }
+}
+
 + (void)leaveBreadcrumbWithBlock:
     (void (^_Nonnull)(BugsnagBreadcrumb *_Nonnull))block {
     if ([self bugsnagStarted]) {
@@ -231,6 +242,19 @@ static NSMutableArray <id<BugsnagPlugin>> *registeredPlugins;
     (NSString *_Nonnull)notificationName {
     if ([self bugsnagStarted]) {
         [self.client crumbleNotification:notificationName];
+    }
+}
+
++ (void)leaveBreadcrumbWithMessage:(NSString *_Nonnull)message
+                          metadata:(NSDictionary *_Nullable)metadata
+                           andType:(BSGBreadcrumbType)type
+{
+    if ([self bugsnagStarted]) {
+        [self leaveBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumbs) {
+            crumbs.message = message;
+            crumbs.metadata = metadata;
+            crumbs.type = type;
+        }];
     }
 }